+++ tests/players/test_human_player.py
@@ -1,7 +1,6 @@
 import unittest
-from unittest import mock
+from unittest.mock import patch, MagicMock
 from app.players.human_player import HumanPlayer
-import io
 from app.main import TicTacToe
 
 
@@ -10,8 +9,8 @@ class TestHumanPlayer(unittest.TestCase):
     def setUp(self):
         self.player = HumanPlayer("X")
     
-    @mock.patch("builtins.print")
-    @mock.patch("builtins.input", side_effect=["123Adithya", "Adithya"])
+    @patch("builtins.print")
+    @patch("builtins.input", side_effect=["123Adithya", "Adithya"])
     def test_set_player_info(self, mock_input, mock_print):
         """
             Assert the expected behaviour for both valid and invalid name
@@ -21,16 +20,16 @@ class TestHumanPlayer(unittest.TestCase):
         self.assertEqual(self.player.name, 'Adithya')
         self.assertTrue(is_done)
 
-    @mock.patch("builtins.print")
-    @mock.patch("builtins.input", side_effect=["1", "2", "3"])
+    @patch("builtins.print")
+    @patch("builtins.input", side_effect=["1", "2", "3"])
     def test_get_choice(self, mock_input, mock_print):
         """
             Verify get_choice prints error message for wrong choice
         """
         self.player.name = "Adithya"
         mock_game_obj = TicTacToe()
-        mock_game_obj.is_valid_cell_no = mock.MagicMock(name="us_valid_cell_no")
-        mock_game_obj.is_cell_available = mock.MagicMock(name="is_cell_availble")
+        mock_game_obj.is_valid_cell_no = MagicMock(name="us_valid_cell_no")
+        mock_game_obj.is_cell_available = MagicMock(name="is_cell_availble")
         mock_game_obj.is_valid_cell_no.side_effect = self.cell_no_side_effect
         mock_game_obj.is_cell_available.side_effect = self.cell_available_side_effect
         position = self.player.get_choice(mock_game_obj)