+++ changelog.md
@@ -4,10 +4,23 @@ Version
 ### Added
 - Simple command-line TicTacToe game without bot
 
+
 ## V2
 
 ### Added
 - Enabled Bot player
+- Added MiniMax algorithm for bot
+- Added trigger to ask opponent player's option
 
 ### Fixes
 - Fix on Game status function to detect left cross connection as a win
+
+## V3
+
+### Added
+- Code Refactored
+- Fixed Linters
+- Bug fixes on game movement
+- Improved Game board display
+- Enabled UnitTests for all the class
+- Improved Version of documents